iio: adc: ad7124: spi communication fix
	Before this patch, the ad7124 requested from the
device 4 bytes when performing a read operation
on the the data register. The data register of ad7124 is
only 3 bytes. This puts the chip's serial interface in wrong
logic state and the driver receives wrong data.

	This is generated by the fact that ad_sigma_delta
library computes the size of the data register using attributes
from iio_chan_spec of channel 0: .shift + .realbits.
Setting the shift to 0 and realbits to 24 in iio_chan_spec will
make ad_sigma_delta to compute the right size for the data register.

	The storage set to 32 bits in iio_chan_spec causes
a misalignment in the triggered buffer. This causes iio-osciloscope
to show a distorted waveform. In order to correct this, the
trigger handler from ad_sigma_delta is overridden with another
one similar which realligns each sample correctly in the buffer.

Signed-off-by: Alexandru Tachici <alexandru.tachici@analog.com>